+++ dashboard/src/scenes/auth/index.js
@@ -1,16 +1,17 @@
-import React from "react";
-import { Route, Switch } from "react-router-dom";
+import React from 'react';
+import { Switch } from 'react-router-dom';
 
-import Reset from "./reset";
-import Forgot from "./forgot";
-import SignIn from "./signin";
+import Reset from './reset';
+import Forgot from './forgot';
+import SignIn from './signin';
+import SentryRoute from '../../components/Sentryroute';
 
 const Router = () => {
   return (
     <Switch>
-      <Route path="/auth/reset" component={Reset} />
-      <Route path="/auth/forgot" component={Forgot} />
-      <Route path="/auth" component={SignIn} />
+      <SentryRoute path="/auth/reset" component={Reset} />
+      <SentryRoute path="/auth/forgot" component={Forgot} />
+      <SentryRoute path="/auth" component={SignIn} />
     </Switch>
   );
 };